requirements.txt
pandas
torch
transformers
accelerate
python-dotenv
sentencepiece
protobuf